+++ src/test/php/com/mongodb/unittest/ScramSHA1Test.class.php
@@ -70,7 +70,6 @@ class ScramSHA1Test {
       [
         'saslContinue'   => 1,
         'payload'        => '',
-        'done'           => true,
         'conversationId' => 1,
         '$db'            => self::DATABASE
       ],